# packages/contracts/openapi/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AlgoMinutes API",
    "version": "1.0.0-v1",
    "description": "The AlgoMinutes API contract, generated from the zod schemas in @algominutes/contracts. Do not hand-edit the emitted JSON — edit the schemas and regenerate."
  },
  "servers": [
    {
      "url": "{origin}",
      "variables": {
        "origin": {
          "default": "https://api.algominutes.app"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Firebase Auth ID token. Absent only on the public routes: POST /v1/shares/read and POST /v1/client-error."
      }
    },
    "schemas": {
...
      "IsoDateTime": {
        "type": "string",
        "format": "date-time",
        "example": "2026-08-15T09:41:07.123Z"
      },
...
      "RedactionMeta": {
        "type": "object",
        "properties": {
          "applied": {
            "type": "boolean"
          },
          "scheme": {
            "type": "string"
          }
        },
        "required": [
          "applied",
          "scheme"
        ],
        "example": {
          "applied": true,
          "scheme": "shared/redaction.cjs"
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "Note not found"
        }
      },
...
      "SharedTranscriptLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "speakerTag": {
            "type": "integer",
            "nullable": true
          },
          "speakerName": {
            "type": "string",
            "nullable": true
          },
          "startMs": {
            "type": "number",
            "nullable": true
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "speakerTag",
          "speakerName",
          "startMs",
          "text"
        ]
      },
...
      "SharedNoteRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "SharedSummary": {
        "type": "object",
        "properties": {
          "gist": {
            "type": "string",
            "nullable": true
          },
          "actionItems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keyDecisions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "gist",
          "actionItems",
          "keyDecisions"
        ]
      },
      "SharedNoteResponse": {
        "type": "object",
        "properties": {
          "note": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "nullable": true
              },
              "createdAt": {
                "$ref": "#/components/schemas/IsoDateTime"
              },
              "scope": {
                "type": "string"
              }
            },
            "required": [
              "title",
              "createdAt",
              "scope"
            ]
          },
          "summary": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SharedSummary"
              },
              {
                "nullable": true
              }
            ]
          },
          "transcript": {
            "type": "object",
            "nullable": true,
            "properties": {
              "lines": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SharedTranscriptLine"
                }
              },
              "truncated": {
                "type": "boolean"
              }
            },
            "required": [
              "lines",
              "truncated"
            ]
          },
          "expiresAt": {
            "$ref": "#/components/schemas/IsoDateTime"
          },
          "redaction": {
            "$ref": "#/components/schemas/RedactionMeta"
          }
        },
        "required": [
          "note",
          "summary",
          "transcript",
          "expiresAt",
          "redaction"
        ]
      },
...
      "ClientVersion": {
        "type": "string",
        "example": "ios/1.4.0"
      },
...
    },
    "parameters": {
      "ClientVersion": {
        "schema": {
          "$ref": "#/components/schemas/ClientVersion"
        },
        "required": true,
        "name": "X-AlgoMinutes-Client",
        "in": "header",
        "description": "Client build identifier, \"<platform>/<semver>\" e.g. ios/1.4.0. The server refuses unsupported builds with 426."
      }
    }
  },
  "paths": {
...
    "/v1/shares/read": {
      "post": {
        "summary": "PUBLIC read of a shared note. No bearer token — the link token is the credential.",
        "tags": [
          "share"
        ],
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/ClientVersion"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharedNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Redacted share view.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedNoteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found (indistinguishable from expired/revoked/malformed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests (per-IP limiter).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}